Create README.md
The goal of this project was to make an efficient tool that can search through an XML file for given IDs and replace specific values, or create new nodes that are related to the given ID. This program construct a Document Object Model (DOM) by reading in an existing XML file and manipulating different values between tag names. I have used the Document Object Model APIs in order to modify, rearrange, delete, and create nodes in the existing XML file.